# protocol/openapi.yaml
openapi: 3.1.0
info:
  title: Open Pigeon Protocol HTTP API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
  description: Public HTTP contract for compatible Pigeon chat servers.
servers:
  - url: https://api.pigeonsms.aldi.best
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /spaces/{spaceId}/migration:
    get:
      tags: [Portability]
      operationId: exportSpaceMigration
      parameters:
        - $ref: '#/components/parameters/SpaceId'
      responses:
        '200': { description: Portable pigeon-migration bundle }
  /spaces/migrate:
    post:
      tags: [Portability]
      operationId: importSpaceMigration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bundle]
              properties:
                bundle: { type: object }
                name: { type: string, maxLength: 48 }
                force: { type: boolean, default: false }
      responses:
        '201': { description: Imported nest and media-copy status }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: pigeon_session
  parameters:
...
    SpaceId:
      name: spaceId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Id' }
  schemas:
    Id:
      type: string
      pattern: '^[0-9]+$'
...